+++ src/c_orm_query_fluent.c
@@ -1034,8 +1034,8 @@ c_orm_query_eager_load_impl(c_orm_query_t *q, const c_orm_table_meta_t *meta,
         *p++ = ' ';
       }
       first = 0;
-      w = C_ORM_SPRINTF(p, 4096 - (p - columns), "%s.%s", meta->name,
-                        meta->columns[col_i].name);
+      w = C_ORM_SPRINTF(p, (size_t)(4096 - (size_t)(p - columns)), "%s.%s",
+                        meta->name, meta->columns[col_i].name);
       p += w;
     }
 
@@ -1045,8 +1045,8 @@ c_orm_query_eager_load_impl(c_orm_query_t *q, const c_orm_table_meta_t *meta,
         *p++ = ' ';
       }
       first = 0;
-      w = C_ORM_SPRINTF(p, 4096 - (p - columns), "%s.%s AS %s_%s",
-                        rel->target_meta->name,
+      w = C_ORM_SPRINTF(p, (size_t)(4096 - (size_t)(p - columns)),
+                        "%s.%s AS %s_%s", rel->target_meta->name,
                         rel->target_meta->columns[col_i].name, relation_name,
                         rel->target_meta->columns[col_i].name);
       p += w;
